+++ docs/changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.2
+^^^^^
+* New ordering algorithms square, horizontal and vertical.
+* New command line argument ``--ordering``.
+* Old algorithms maxside, width, height and area are now orderings.
+
+
 0.1.9
 ^^^^^
 * New command line argument ``-z``, ``--no-size`` to avoid adding the image width and height to the sprite.